+++ tomcat/src/main/java/com/techcourse/controller/SignupController.java
@@ -3,10 +3,6 @@ package com.techcourse.controller;
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 import java.io.IOException;
-import java.net.URISyntaxException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 import org.apache.coyote.http11.HttpResponseStatus;
@@ -30,19 +26,13 @@ public class SignupController extends AbstractController {
             return;
         }
 
+        handleSignup(response, account, password, email);
+    }
+
+    private static void handleSignup(HttpResponse response, String account, String password, String email)
+            throws IOException {
         User user = new User(account, password, email);
         InMemoryUserRepository.save(user);
         response.sendRedirect(HttpResponseStatus.FOUND, "/index.html");
-    }
-
-    private void serveStaticFile(String path, HttpResponse response, String contentType) throws IOException, URISyntaxException {
-        final var resource = getClass().getClassLoader().getResource("static" + path);
-        if (resource != null) {
-            final Path resourcePath = Paths.get(resource.toURI());
-            byte[] body = Files.readAllBytes(resourcePath);
-            response.addHeader("Content-Type", contentType);
-            response.setBody(body);
-        }
-        response.send();
     }
 }
