+++ tomcat/src/main/java/com/techcourse/controller/StaticController.java
@@ -1,11 +1,6 @@
 package com.techcourse.controller;
 
-import java.io.IOException;
-import java.net.URISyntaxException;
 import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 
@@ -29,17 +24,6 @@ public class StaticController extends AbstractController{
 
         final byte[] body = "Hello world!".getBytes(StandardCharsets.UTF_8);
         response.setBody(body);
-        response.send();
-    }
-
-    private void serveStaticFile(String path, HttpResponse response, String contentType) throws IOException, URISyntaxException {
-        final var resource = getClass().getClassLoader().getResource("static" + path);
-        if (resource != null) {
-            final Path resourcePath = Paths.get(resource.toURI());
-            byte[] body = Files.readAllBytes(resourcePath);
-            response.addHeader("Content-Type", contentType);
-            response.setBody(body);
-        }
         response.send();
     }
 }
